refactor: List -> Catalog 이름 수정
